cut extra tof values if less int values are present
Some acqu files report a greater number of tof/mass values than the number of
intensity values in the fid files. Now the last tof/mass values are removed to
keep the number of tof/mass/intensity equal and a warning is thrown.
Thanks to Frederic Foucault <frederic.foucault@mabritec.com> for reporting
this problem.